refactor: update delete functionality for groups and associated journals
Refactored the delete functionality for groups in the GroupManager component. Updated the code to delete associated journals from all journal collections (standard_journals, advanced_journals, and top_journals) when a group is deleted. This ensures that all related data is properly removed from the database. Also invalidated the "groups", "students", and "journals" queries in the onSuccess callback to reflect the changes.